Распределение кода по модулям
